ThreadPoolExecutor now survives a fork.
